requirements.txt
flask
numpy
scikit-learn
requests
gunicorn
beautifulsoup4
python-whois
xgboost